requirements.txt
langchain
langgraph
langchain-ollama
langchain-chroma
langchain-openrouter
uuid
python-dotenv